Initial commit. Defined Tensor, Vector classes
This is my first commit for this project. I began by defining a vector and tensor classes. These classes, as their names would imply, are intended to act like vectors and second order tensors, respectivly.

Vector class: The vector class has been defined to store a 3 element array (since all vectors that I will be working with are defined in 3d space).  I defined 3 constructors for this class, a default one, an array based one, and a component based one. The default one initializes the member array to all zeros. The array one accepts a 3 component array as an argument and then uses its components to define the components of the vector's array. The component version accepts the 3 components are separate double argumnets. Additionally, I endowed the vector class with a number of operator overloads. I have defined addition, subtraction, equality to array's, (), and [] for vector classes.+, - work as expected, as does equality with an array. The () and [] are used to access elements of the array. These operators return a REFERENCE to an element of the array in the vector. This allows you to directly modify components of the vector in an intuitive manner. Finally, I added a print method which prints the components of the vector.

Tensor class: I defined a second order tensor class. This class contains a 9 element array (which stores the 9 components of the tensor). The class includes two constrcutors, a defult one and a component based one. The class also contains operator overloads for addition, multiplication with other tensors, and multiplication with a vector. The addition uses component wise matrix addition. The tensor multiplication uses matrix-matrix multiplication and the multiplication with vectors uses matrix-vector multiplication. Finally, I included a print function which displays the 9 components of the tensor.

I also wrote a Dyadic product function which accepts two vectors and returns their dyadic product (as a 2nd order tensor). The main function currently contains test code/will be updated as progress continues

I plan on moving the class definitions for the Vector and Tensor classes into header/source files asap. The main file is getting wayyyyy too cluttered.