Improved Vector, Tensor classes
I began by adding a uByte (and Byte) type definition. I'm not sure if I'll use these in the future, though I expec them to be useful (I've used them in many other projects).

To begin, added the const keyword to as many methods as I could. It was added to all argument and method prototypes that could use it. This will help prevent any accidential modifications to member variables/better communicates what these methods should NOT modify.

I added another = operator, as well as some new () and [] operators for the vector class. The new = operator is used to assign a vector to another vector (previous one assigned it to an array). this is very useful for initialization. Technically this overloads a method that the compiler would have defined otherwise. The new () and [] operators are used for in a const context (doesn't modify member variables of the object that uses the operator). not sure why this is needed, but it was (need to look into this).

I added a similar assortment of operator overloads for the tensor class, though I choose to not include a [] operator for the tensor class. [] can only accept one argument ( () can accept any number). This makes it impossible to use matrix subscripts with (). Now, matrix A's i,j element can be accessed with A(i,j).

Finally, I modified some of the existing main function code to test the new functionality.