Created separate tensor, vector files
The main file was getting too cluttered. To fix this, I moved my definitions for the Tensor and Vector classes into separate files. Both classes were given a header file (with the class definition) and a source file (with the method definitions)

The header and source files were then included in the main file.