Added += overload, removed a vector constructor,
I added overloading for the += operator for the Vector and Tensor classes. This operator will be used extensivly in the SPH algorith.

I also removed the 'vector based' vector constructor. This constructor worked by accepting an array as a constructor argumnet. I think that this constructor was obselete... it looks better to use the = overload.  As a result, I removed this constructor (probably didn't need to remove it, but I removed it just in case my interface changes/I don't want to maintain that constrcutor).

I also added a few more comments here and there as well as touched up spacing... all to make the code look a little nicer