Added Tensor *, /, *= Inverse
I began by adding operator overloading for scalar multiplication and division. Let A be a tensor and c be a scalar. I defined the operations A*c and c*A. I also added an operator to divide a tensor by a scalar, A/c (Note: c/A is nonsense!). The c*A operator had to be added as a friend function (since the operator is defined for a double, not for a tensor type object).

I also added overloading for the *= operator with a scalar. Thus, A *=c; is now defined!

Finally, I also added an inverse function for the tensor. This retruns the inverse of the tensor that calls it. After some debugging, I believe that this function is working.