Started Particle class, added magnitude method for vectors
First, I added a magnitud method for the vector class. This method, as the name would imply, returns the magnitude of the vector that calls it.

I also began writing a particle class. I'm a little conflicted about what should go into this class. A particle fundamentally needs a reference and spacial position (x and X) as well as an acceleration. It also makes sense to store the force applied to that Particle, and its volume.

However, It's unclear if things like the deformation gradient F, shape tensor A, or the Piolia-Kirchoff tensors P and S should be stored in the particle. Further, it's inclear if I should include a particle's neighbor list. The reason is that the elements in this list will be array indicies... this means that the particle can't function unless an array of such particles has already been defined. This seems to ruin the idea that the particles are objects/feels weird to me.

I think that it makes some sense to include the shape matrix, deformation gradient, P, and S in each particle since we calculate these quantities for each particle. Grad_W, on the other hand, is calculated several times for each particle and is really just used to calculate other quantities (such as A). Therefore, no W member is included.

I also included private functions to calculate Grad_W(Rj,h), update A, update F, update P, and update the force applied to the particle, and  update the acceleration of the particle. The idea here is that a particle always stores its previous acceleration, as well as its current position.

When the method Update_x is called, the function uses the current position to update A, F, P, Force, and Acceleration (using the private methods). The new acceleration is then used to update the position (x), at which point the function returns. Each call of Update_x basically functions like a time step.

I'm a little uncertian abut this setup, and am especially unsure about how to generate the neighbor list.