Made some updates to the Particles class
I added a copy constructor and destructor to the particles class. It should be noted that objects of one class can access private data of the objects of another class. Thus, my copy constructor directly copies elements from P_In to the local members.

I also added a static 'Num_Particles' member to the static class. This will be used to keep track of the total number of particles created in a program. All particle constructors now increase the number of particles by 1, the destructor decreases it by 1.