Updated (and maybe finished) copy constructor + destructor for particle class
I updated the copy constructor of the particle class. The copy constructor now performs a deep copy of P_In. This means that the new Particle gets the same contents as the origional particle, but the dynamically allocate arrays have new memory locations.

To understand why we need to do this, let's consider what would happen if we set the Neighbor_list and Grad_W_Tilde pointers in the new particle to the addresses in P_In. The copy constructor is used when passing an object to a non-reference argument of a function and when a returning non-reference object. These objects are generally temporary. When a function returns, any objects that were created when accepting arguments are deleted. This would therefore free the memory in the origional particle. Thus, after the temporary copy was deleted, the origional particle would no longer have a Neighbor_List or Grad_W_Tilde member. This is clearly a problem. Therefore, to fix this, we allocate new memory in the copy constructor. The contents of the origional dynamic array are then copied into the new dynamic array in a element-by-element manner. This way, when a temporary copy is deleted, it free's the new dynamic array, not the origional one.

I also added delete [] statements to the destructor to free the two pointer members.