Redid Particles class + Added det method to Tensors
I basiclally redid the entire Particle class. For one, many of the private 'Update' methods and private members (such as force, accleration, F, etc...) have been removed. The removed variables will instead be local variables in the new Update_Particle_Position friend function. Likewise, the private methods have been moved into this function (or will be). this was done to reduce the memory footprint of each particle

The new Update_Particle_Position function has been partially defined. This function, as the name implies, update's the particle's position. The idea here is that it takes in a particle + the particles array (from the main function). This is used to calculate the force + acceleration on the particle, which is then used to find the new position. The new position is then stored in a 'x_new' member variable. The reason for this is that we need to know the particle's old positions when calculating the force. If we updated a given particle's x vector when we updated its position, then any neighbors of that particle would see it's new position rather than the position it had at the beginning of the time step. This leads to indeterminsitic results (results depend on the order you update the particles in). Thus, to avoid this issue, the new position is stored in a seperate vector. After each particle's position has been updated, x_new is copied to x (for each particle). This concludes the time step.

The Calc_Inverse function has been moved into the Set_Neighbors function (this is the only time you'd run the A_Inv method anyway, so it made sense to move it into the Set_Neighbors function).

I also added a deep copy = operator overload. This operator overloads the implicit shallow copy = operator. This was done to eliminate the risk of double freeing/bad stuff happening with the dyamic members of the particle. The = operator method looks identical to the deep copy constructor method, except that it returns *this at the end.

Finally, I added a determinant method to the Tensor class. After studying the algorythm in more detail, it appears that I will need to calculate the tensor determinant. I choose to calculate the det in this function rather than storing the det as a member because the tensor components may change with time. In theory, I could update the det each time that a component of the tensor is changed/store the determinant as a member... this may work better (I will need to test once the code is fully operational). A test was added to the Tests.c source file to test the new determinant function (and it appears to be working!)

Finally, the TODO document has been updated to reflect progress.