Cleared out syntax errors in Particle class, added a few functions
I finished the Paritlce class functions, though I was not able to calculate new position (just got to acceleration).

I also need to add to add the scalar A when calculating W and Grad_W.

I added a calc_W method to the particle class, this calculates W for a given Rj (neeeded to calculate hour class force).

I also initialized the static member variables for the Particle class just so that the code would compile... this needs to be removed/replaced with real values ASAP.

I also added a - operator overload to the Tensor class (needed to calculate forces).

I somewhat updated the TODO file, though this needs more work.