Added a 2 way linked list class + Tests
Now that the particle class is finished (but not debugged), I am working twoards a neighbor search. I decided to do this using a 2-way linked list. I doubt that I'll need to be able to work both ways, but it was a fun exercise to implement this structure.

The list consists of a series of nodes. A node is a struct that is private to the List class. Each node contains a PArticle_Id and two pointers, one to the previous node in the list and one to the next node in the list. Each list contains 3 things: a pointer to the start of the list, a pointer to the end of the list, and the total number of nodes in the list.

Methods have been defined to add or remove elements to the start or end of the the list. The add functions accept an integer and use this as the Particle_Id for the appended node. The remove functions return the value stored in the removed node.

Since memory is dynamically allocated, care has been taken to perform deep copies in the copy constructor and = operators (this is why I included explicit definitions for these methods, even though I doubt I'll ever use them).

I also added tests to ensure that the list is working. This revealed a few bugs that have since been fixed. I believe the list class is working.

The todo file has been updated to reflect progress.