Updated LeapFrog method, removed incompressibility condition, cleaned up Particle class
I learned that the calculated acceleration is assumed to apply at the new position. Thus, in the ith step we are using x_i to calculate x_i+1 using the acceleration a_(i+1) which is what we're calculating. Because of this, the acceleration always applies to the new position (will apply to a full time step ahead of starting position). Because of this, It made sense to implement the conventional leap frog method.

Also, since we are not working with plane stress, we can't use the p(J-1) penality term in the strain energy function (this is thrown in to ensure incomrpessibility). With a 3d case, we need to use a different incompresibility condition. I will try implementing eq 6.138 in the H book.

I also cleaned up the particle class a bit (explaining what the static members do) by reorganizing the static members + renaming a few (to better names).

Todo file has also been updated.