Implemented particle tests, fixed numerous logical errors in particle class.
I managed to implement the 'particle test'. This test consists of creating a 3x3x3 cube of particles. The particles are stored in a 3d array, and their initial position is simply their coordinates in the cube. All particles, except for one, start at rest. I give the (1,1,0) particle an initial velocity of <0,0,5> (twoards the center). Once the particles are given these paramaters (as well as their mass and volume), the particles print their parameters (to  make sure that the setup worked). Next, we assign the neighbor list. This is done using a linked list (see code for implementation). I plan on moving this neighbor search to a dedicated function in the near future. Once the neighbors have been found and set, the particle data is again printed (this time with neighbor data). Finally, I compute one time step and then print the new particle data.

Writing this test revealed several logical errors in my particle methods. First off, I realized that I needed a method to determine if two particles are neighbors. I wrote a friend function (in the particle class) to do just this.

First, I  realized that the particle destructor caused major issues if the neighbors had not been set. Before the neighbors are set, the pointer members point to nothing. Thus, when the destructor is called, it tries freeing memory that isn't allocated, this caused the program to crash. To fix this, I only run the delete calls (in the particle destructor) if the nehgibhrs have been set.

Next, I found a bug at the start of the Set_Neighbors function... if the passed N value is 0 (implying 0 neighbors) then the function crashes (tries allocating 0 bytes of memory, causes terrible bugs). To fix this, I test if N == 0 at the start of the function. if it is, then I return

Next, I removed the P_Neighbor particle from the set_neighbors method (just to conserve memory/it wasn't needed).

I also found a big with RJ being zero... this causes the grad function to blow up, iterally... not sure how to fix this (need to discuss).... I'm thinking of putting in a check to see if MAg_rj = 0, then set Grad_W[j] = 0. this forced me to prevent a particle from being its own neighbor (if so then Rj = 0, but |rj| = 0... Rj/|rj| = undefined).

there was a sneaky little bug in the set_neighbors function. In this function, I had declared a local tensor A. This had the same name as the scalar A (shape function amplitude). When calculating Grad_W, I wanted to multiply some constants by Rj and A (scalar). The issue was, the program was using the local A (tensor) instead of the global A (scalar). Both operations result in a vector. Because of this, the code wasn't returning any errors, though the result was wildly wrong. To fix this, I choose to rename A (scalar) to Shape_Function_Amp. This name explains itself and removes any ambuguity.

I also found an issue with the = operator... this operator wasn't allocating (via a deep copy) the r and R vectors... this caused major issues whenever a particle that was created using the = operator (like P_Object in the update pos class) was deleted, since it's R, r were not allocated but its neighbors were set, causing the delete [] r and delete [] R calls to hang. This was fixed by adding deep copying for R and r to the = operator.

I also found a few bugs in the update P function... for one, the I1 (line 302) was using C(3,3), which is out of bounds... I was using math indicies, not cs indicies... correcting the indicies fixed this issue.

I also added a print method to the particle class, this is used for debugging.

the new test code was added to the tests header/source.

The todo class has also been updated to reflect progress/new goals.