Added Tensor [] access to improve runtime.
I discovered a few things regarding the compiler, and how fast I can make things run.

I quickly realized that using the (row, col) method to access elements of a tensor is slow. This method actually performs 5 artihemtic operations every time that it's called. The (row, col) method for accessing tensor components is good for a public interface, but too slow for things in the tensor class's scope. To fix this, I defined a new access method: []. This is a privae method. The idea is that all tensor methods will access elements in this way, while the (row, col) access scheme will be used by the public interface. The [] method accepts one index (the array index), runs no checks, and simply returns T[N] (where N is the input index). This method can EASILY cause a segmentation fault (if the input is negative, or >= 9). As a result, this method has been made expressly private.

With this new method, I redid the tensor-tensor product and compount tensor-tensor product methods. The result is code that runs 30% faster!

I also updated where the timers are placed, and the print statements in the Timing_Tests functions (int test.c)

I now need to update the other methods, eliminating all loops and all uses of the (row, col) access scheme in private methods!