Removed 0 initialization of vectors, tensors. Vector [] private, updated friends
I began by getting rid of the 0 initialization of the vector and tensor classes (the default constructor initialized all elements of T to zero).  This change makes it dangerous to work with tensor and vector objects before they have been initialized. To counter this, I explicitly initialized the x, X and vel particle members to the zero vector (I'm leaving it up to the user to define P, F).

Making this change, however, substantially improved performance. The time to calculate 10,000,000 tensor-tensor products decreased from ~.97 seconds to ~.85 seconds (best case). Likewise, the time to perform compound compound tensor-tensor multiplication droped from ~.95 seconds to ~.83 seconds. Finally, the time to compute 10,000,000 tensor inverses dropped to around .72 seconds (from .89 before).

Next, I moved to optimize the vector class. All of the would-be loops in the vector class had already been unrolled. However, the component access method, [],. still ran checks. I moved the [] method to the private section of the vector class and removed all checks (I also made the function inline, so its definition can be found in vector.h). This should improve performnace. All vector methods now use this private, but faster, access scheme. The () access scheme is reserved for the public interface (it still runs checks).

I also updated the dyadic product function. It now uses the [] access scheme for tensors. However, this change along with making [] private for vector, created a bit of an issue. The dyadic product no longer had access to the [] method for the vector or tensor classes. I made the dyadic product function a friend of both classes, but this created another issue. The dyadic product funtion accepts two vectors and returns a tensor. To be able to use it, you need to know what a vector and a tensor is. This created a big issue. The vector class, which is included before the tensor class, can't see the tensor class. To fix this, I created a new file, the 'classes' header file. This headerfile contains class prototypes for all 4 classes in my code. This fixed the problem.

However, making [] private for vectors created another problem in the tensor class, I could no longer use this access sceheme for tensor-vector multiplication and the sort. To fix this, I made the Tensor class a friend of the Vector class.

I also noticed that the magnitude, vector dot product, and print functions were not 'friends' of the vector class. This made it impossible for them to use the [] access method. To fix this, I made these functions friends.

With these changes, I updated the tests (removing the default constructor and [] access tests).

I also updated the compiler guard definitions (removed the _ from the front). This was done because C++ uses definitions with an underscore followed by a capital letter to define some of its internal variables. I removed the _ from my definitions to avoid potential conflict.

Finally, I updated the todo file to reflect progress.