Fixed a nasty tensor-tensor product bug
I fixed a nasty bug associated with tensor-tensor multiplication. The old code used the += operator for all 27 statements in the tensor-tensor product code. This was fine when the tensors were initialized to zero. However, now that they are not zero-initialized, using += rather than = when the elements of the tensor are first updated results in nasty runtime behavior. I fixed this by replacing += with just = in the 9 statements that first update the 9 elements of T (each element of T is updated 3 times, I only replaced the statements that correpond to the first update). This change fixed the bug.