Improved varialbe naming
This commit mostl focuses on updating some variable names. I updated names in the List, and Particle classes.

List naming changes: I realized that the list variable/method names were quite confusing. When talking about the list, it makes sense to discuss the 'beginning' and the 'end'. However, when talking about appending elements, it makese sense to discuss appending to the 'front' or 'back' (rather than appending to the beginning or end, which sounds more confusing). To make the naming scheme more logical, I changed the List pointer names from 'Front' and 'End' to 'Beginning' and 'End'. the method names were changed from 'Add/Remove_Front' and 'Add/Remove_End' to 'Add/Remove_Front' and 'Add/Remove_Back'. I believe this naming scheme makes it more inutitive to work with list objects. I updated the List.c (and test) files to this new naming scheme.

For the particle class, the name 'Neigbor_list' was confusing, since it suggests that the object is a List, when it's actually an array. To remove this confusion, I changed the variable's name to 'Neighbor_IDs' which more accuretly describes what it is.

I also redid a lot of the variables in the particle class methods. I tried to eliminate the usage of stuff like 'P_In.R[j]' by instead defining a variable that points to P_In's R vector, then using that variable to access P_In's elements. This may add a tiny bit of overhead, but it makes the code far more readible. Plus, I'm pretty sure the compiler will take care of this unecessairry (but helpful) aliasing/make the code run just as quick.  These kind of changes were made throughout the Paricle class.

I also changed the Vector/Tensor classes so that their print functions prints in sci notation (formatted to be 9 characters wide with 2 deimal places). This helps when debugging the particle position update methods.