Updated Strain energy function + Added a ^ operator for tensors
I updated the strain energy function using the equation that I derived yesterday (for 3d objects). With this change, the particles (mostly) stay put/do not move when in a stationary configuration.

I removed the *=(double) method for the Tensor class. This was done because I was literally never using this operator/it was just adding unecessairry maintence.

I also added a ^ operator overload for the Tensor class. This operator is used to calculate the Inverse, Transpose, and Inverse Transpose of a given tensor.

To understand how this works, consider a tensor called A. If the user passes T to the ^ operator (A^T) then it returns the transpose of A. If the user passes -1 to the ^ operator (A^-1)if they pass -1 then the ^ operator then it returns the inverse of A. Finally, if the user passes -T to the ^ operator (A^-T) then it returns the inverse transpose of A.

So how does this work and what is T? To make this work, I established a definition for 'T' in the main file. This is simply equal to 2.  (#define T 2). Thus, whenever the user types T, the #define directive replaces it with a 2. Likewise, whenever the user types -T, the #define directive replaces it with -2. The actual ^ operator itself is setup to accept an integer argument. If that argument is -1, it returns the inverse; if the argument is 2, then it returns the Transpose, and if the argument is -2 then it returns the inverse transpose.

Though this new feature is, in my humble opinion, quite spiffy, it does come with some drawbacks. For one, since the letter 'T' is now being used as an alias for 2, you can't declare a variable named T ANYWHERE in the code. This created some serious problems, given that the Matrix representation of my Tensor class was called T. To fix this, all Tensor type variables have been changed to S, rather than T (T_In becomes S_In, det_T becomes det_S, T array becomes the S array). This is acually where most of the changes in this commit came from.

I also added in tests to tests the new methods/operators.

I debated removing the Inverse and transpose friend functions and making the .Inverse and .Transpose methods private. However, I decided against this. There's no reason to prevent users from using the .Inverse or .transpose methods, even if they are less spiffy.

Finally, the use of ^ operator can cause some SERIOUS trouble if not used properly. the ^ operator has pretty low precedence. Therefore, something like A^-1*B will actually work like A^(-1*B), which will return an error. The correct syntax is (A^-1)*B. the same applies for multiplication of tensors. For example, A*B^-1 will run as (A*B)^-1. If the desired output is A times B inverse then you should type A*(B^-1). In general, you should ALWAYS wrap any use of the ^ operator in a set of parenthesis. This will guarantee that that the code runs properly/doesn't do weird stuff.

I also moved my dyadic product identity test to the Tensor_Tests function (in Tests.c). The tensor tests have also been updated to include blank lines between test printouts (it just makes the final output a little nicer to look at).

Finally, the TODO file has been updated.