Started messing around with printing to file
I wrote some code that attempts to print some results to a file.

This proved to be quite tricky. It turns out that C kinda sucks at editing files. There is no way to, for example, delete a character or line from a file. I want my file to contain the number of iterations in addition to the number of particles. The best solution I could come up with (have not yet implemented) is to keep the position data in one file and the paramaters (like number of iterations + number of particles) in another file. The idea here is that I can write to the position file each cycle, then only update the paramaters file at the end. I still need to implement this/read it in from matlab.