Added units to particles, reordered methods, updated tests
I added units to all statements in the particle (and test) classes that use units. For example, if a statement is used to calculate P, I added a comment saying Mpa (since P's components are in units of Mpa). I used a standard for this.. each unit comment starts with a colon at character 90 followed by a single space and then the unit. For example ': mm'.

I also reordered the methods on several classes. I moved the = operator up by the constructors/destructor in all four classes. This change was made in both the header file and source file.  I also moved all of the Particle class's 'setting functions' except for Set_Neighbors into the class header.

I also added some comments to the List header/source. These comments mostly just explain what some of the methods do.

I eliminated the r pointer and density static member from the particle class. I added in a w pointer. I relized that it didn't make sense to store the neighbor's positions in the class. Doing this required updating the entire array after each step. I believe it will be faster to instead get the desired particle's position (we know it's ID in the particle array using the neighbor ID array... so we can get it's position that way) and calculate rj using that ( Particles[neighbor_ID].x - P_In.x). I also realized that it was silly to call a function to calculate W. W is a lagrangian quantity, so it makes sense to calculate it at the start, as soon as the neighbors are known. I modified the code do do this by adding a W pointer (which gets dynamically allocated to an array of Num_Neighbors elements) and removing the Calc_W method. This change, interstingly, improved performance by almost 20%!

Switching to actual units forced me to make a few changes in the code. These changes inclde changing the initialization values for the static class members and multiplying the acceleration by a million (something I'm skeptical about, but the units check out).

I added tests to run 1000 time steps for a 10x10x10 cube of particles. One face of this cube is pushed at a fixed velocity. The initial and final particle positions are saved to a file. I plan on updating the file output to adhere to the system that was emailed to me (vtk thing). I ran this code under a number of differnet viscosity, E, and dt values. A lot lead to divergence. I need to figure out better values for these constants. The setup that I currently have diverges. The hour glass force appears to be the culprit... I need to fix this.

Finally, the todo file has been updated.