Print to VTK format, fixed some bugs
I completely redid my printing-to-file function to print to the desired VTK format. In the new system, particle data is printed every 100 steps to a named file. These files are all stored in a sub-directory named 'Particle_Positions'.

Making thsi change required a number of helper functions. Most of these helper functions are used to generate the file name, rather than the file contents.

I choose to place the VTK_file functions and variables inside of a namespace. I did this mostly to experiment with namespaces.  This also has the advantage of 'hiding' the functions and variables to everything in the code that does not use the namespace. This is important in this case because many of the VTK file memebers are global variables. Thus, any unintentional modifications to these global variables would ruin everything. Placing the functions/variables in a namespace allows me to restrict which parts of my code can see/run those methods.

I created a new source/header file to house the new VTK file namespace. functions/header.

 I also included the relevant headerfiles in each of the source files. These new include statements don't do anything, I just think that help communicate which classes/files are used by which source files.

I also caught a bug in the Particle class. The class didn't multiply (h-|Rj|)^3 by the shape function amplitude when computing the W array. This was causing the code to diverge (by introducing some totally whack shape function values. this was causing the Hour glass force to be much larger than it should have been.

I also changed the order of operations in a few functions in an attempt to make the code run quicker. I also added a few explanations (all in particle.c) in an effort to make the code more understandable.