Changed variable names in Set_Neighbors method (Particle.c)
I changed the name fo a few members in the Set_Neighbors method of the particles class. I chnaged the argument 'Neighbor_ID_List' to 'Neighbor_ID_Array'. This was done because the input is an array, not a list. To avoid confusion with my list class, I changed the name..

I also switched from using N to 'Num_Neighbors' throughout the function (when allocating arrays, looping through neghbors, etc...). This was done because I think it makes more sense to allocate an array with length 'Num_Neighbors' then length 'N', using 'Num_Neighbors' is more self explanatory/makes the code easier to read.