Fixed memory allocation problems
I tried increasing the side length of my cube from 10 to 20 (going from 1000 to 8000 particles). This caused a segmentation fault. I found out the issue was that the size of the particles array was exceeding the size of the stack. To fix this, I dynamically allocate the Particles array (so that it now allocates in the heap). With this change, I was able to get the 20x20x20 cube to work.