Improved file output + Added a debug file printout
I ran a few more tests, several of these ended up diverging.

I am trying to understand what is causig the system to diverge. To do this, I wrote a new set of functions. These functions print out the forces acting on each particle. To get these forced, I had to add force member variables to the particle class (particles don't take up much room already, so this should be fine in terms of memory footprint... adding these 3 forces is like adding an extra tensor to the particle, not too bad). I wrote a set of functions that print these data points to a file.

These functions are included in a new namespace, the 'Partilce_Debugger_File' namespace. Again, I choose to use a namespace to restrict access to global variables related to the debugger, and to restrict access to the functions. I made a new source/headerfile to house these functions.

To make thee functions, I began with the VTK_File functions ( literally coppied the entire VTK_File into the new files and worked from there). I quickly discovered that my whole 'append_digits' function was useless. I could instead pad the current file number with 0's and then place that in a char array that could be added to the file name. Once I realized how much easier it was to do this, I removed the Add_Digits function in the VTK_File namespace.

The Debugger functions print the forces to the sceeen. I made this function a friend of the Vector, Tensor, and Particle classes so that I could gain quick access (use []) for vector/tensor elements, and to gain access to private particle members. The results are printed as a text file, since their purpose is to debug, not visualize.

The debugger functions revaled that the internal force is what's causing the divergence. It is unclear at this time if this is a result of the viscisotiy or stress terms. This should be explored further.

I also significnaltly modified the VTK file functions. They now calculate and print the 6 components of the S and E tensors, as well as J. I choose to do this by copying elements from the particles into arrays, then printing the arrays to the VTK file one at a time. I choose to do this because it has better performnace. I wrote a few paragraphs (see VTK_File.c) explaining why this is better.

Back in the VTK file, I added a function to print a specific paramater to the file. This was done because I'm litterally adding 13 things to the file, copying the 'print' code 13 times is an invitation to errors.

I also gave the VTK functions friend status for the vector/particle classes, this allows me to use the quciker [] access schemes for vectors and tensors + access particle methods without an explicit get function (I removed some get functions as a result of this change).